backend/requirements.txt
celery==4.4.2
channels==2.4.0
channels-redis==2.4.2
Django==3.0.7
djangorestframework==3.12.2
gunicorn==20.0.4
psycopg2-binary==2.8.6
redis==3.4.1
requests==2.23.0
tweepy==3.10.0
uvicorn==0.11.3
whitenoise==5.2.0